Package changes and added vault service.
