--- src/main/java/gov/va/ascent/selenium/utils/PropertiesLoader.java
@@ -1,51 +1,0 @@
-package gov.va.ascent.selenium.utils;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.Enumeration;
-import java.util.Properties;
-
-public class PropertiesLoader {
-	public static void main(String[] args) {
-		PropertiesLoader app = new PropertiesLoader();
-		app.loadBaseProperties();
-	}
-
-	public Properties loadBaseProperties() {
-
-		Properties prop = new Properties();
-		InputStream input = null;
-
-		try {
-
-			String filename = "config/env/base.properties";
-			input = getClass().getClassLoader().getResourceAsStream(filename);
-			if (input == null) {
-				System.out.println("Sorry, unable to find " + filename);
-				return null;
-			}
-
-			prop.load(input);
-
-			Enumeration<?> e = prop.propertyNames();
-			while (e.hasMoreElements()) {
-				String key = (String) e.nextElement();
-				String value = prop.getProperty(key);
-				// System.out.println("Key : " + key + ", Value : " + value);
-			}
-
-		} catch (IOException ex) {
-			ex.printStackTrace();
-		} finally {
-			if (input != null) {
-				try {
-					input.close();
-				} catch (IOException e) {
-					e.printStackTrace();
-				}
-			}
-		}
-		return prop;
-	}
-
-}
